New work flow.
Re-organization so that the pre-processing of images (including
background) subtraction is a separate step from the actual segmentation
of cells. I also added the function to specify a global threshold for
the segmentation instead of computing threshold on a per image basis
using the OTSU thresholding.

Also, I added a functions in `utils.py` to compute the OTSU threshold
from a batch of images. This threshold can be used then as global
threshold in the segmentation for a whole batch of images.