Background subtraction.
I implemented a new method to compute the background. The previous
version (based on openCV medianBlur) was crashing when the size of the
sliding window was greater than ~200 pixels. But I though 200 pixels is
not enough, especially when there are clusters of cells.

I implemented an efficient method based on fast-Fourier transforms.
Local averages can be computed very efficiently through FFT (product of
image FT and kernel FT). I applied this method to the log-transform of
the original image, then re-exponentiate it after the operation. This is
a good estimate of the median.

Works well.